- Improved the behavior of the Status Dock to remove filters from the list when a source is deleted, even if the source remains in the program output.
